Split argument parser runtime compilation units
